✨ Feat: Added count in R34Tags and updated types in R34Posts.

+++ src/R34Sharp/Entities/Posts/R34Posts.cs
@@ -16,7 +16,7 @@ namespace R34Sharp
         /// <summary>
         /// The count of posts present in this collection.
         /// </summary>
-        [XmlIgnore] public int Count => Data == null ? 0 : Data.Length;
+        [XmlIgnore] public ulong Count => Data == null ? 0 : (ulong)Data.Length;
 
         /// <summary>
         /// The offset of current collection.
